+++ packages/opa/src/hooks/types.ts
@@ -3,11 +3,16 @@
  */
 
 import { HTTPClient, RequestInput } from "../lib/http.js";
+import { RetryConfig } from "../lib/retries.js";
+import { SecurityState } from "../lib/security.js";
 
 export type HookContext = {
+  baseURL: string | URL;
   operationID: string;
   oAuth2Scopes?: string[];
   securitySource?: any | (() => Promise<any>);
+  retryConfig: RetryConfig;
+  resolvedSecurity: SecurityState | null;
 };
 
 export type Awaitable<T> = T | Promise<T>;
@@ -96,3 +101,10 @@ export interface Hooks {
   /** Registers a hook to be used by the SDK for the after error event. */
   registerAfterErrorHook(hook: AfterErrorHook): void;
 }
+
+export type Hook =
+  | SDKInitHook
+  | BeforeCreateRequestHook
+  | BeforeRequestHook
+  | AfterSuccessHook
+  | AfterErrorHook;