+++ packages/opa/src/lib/config.ts
@@ -24,11 +24,11 @@ export type SDKOptions = {
   /**
    * Allows overriding the default server used by the SDK
    */
-  serverIdx?: number;
+  serverIdx?: number | undefined;
   /**
    * Allows overriding the default server URL used by the SDK
    */
-  serverURL?: string;
+  serverURL?: string | undefined;
   /**
    * Allows overriding the default retry config used by the SDK
    */
@@ -57,7 +57,7 @@ export function serverURLFromOptions(options: SDKOptions): URL | null {
 export const SDK_METADATA = {
   language: "typescript",
   openapiDocVersion: "0.2.0",
-  sdkVersion: "1.6.0",
-  genVersion: "2.415.7",
-  userAgent: "speakeasy-sdk/typescript 1.6.0 2.415.7 0.2.0 @styra/opa",
+  sdkVersion: "1.7.0",
+  genVersion: "2.522.1",
+  userAgent: "speakeasy-sdk/typescript 1.7.0 2.522.1 0.2.0 @styra/opa",
 } as const;